more tweaks and a file
still working on code for leaderboard to update a single team with new array entry.
Wrote and uploaded bash script to push JSON file to database for QR upload.  It will need to be modified depending on where it is pushing the data.  I added notes to the code about that.